+++ HideWordSolver/ANNA/Poc/forward.c
@@ -1,13 +1,13 @@
 #include "poc.h"
 
 void forward(int nb_letter,
-		double **input, // input_neuron x nb_letter
-		double **hidden, // hidden_neuron x nb_letter
-		double **output, // output_neuron x nb_letter
-		double **w_input, // hidden_neuron x input_neuron
-		double **w_output, // output_neuron x hidden_neuron
-		double *b_input, // hidden_neuron
-		double *b_output) // output_neuron
+		float **input, // input_neuron x nb_letter
+		float **hidden, // hidden_neuron x nb_letter
+		float **output, // output_neuron x nb_letter
+		float **w_input, // hidden_neuron x input_neuron
+		float **w_output, // output_neuron x hidden_neuron
+		float *b_input, // hidden_neuron
+		float *b_output) // output_neuron
 {
     // Apply forward propagation to neurons in the hidden layer
 	poc_matrix_product(hidden_neuron, input_neuron, w_input, input_neuron,
