+++ HideWordSolver/ANNA/Poc/poc.h
@@ -15,50 +15,50 @@
 #define nb_letters 4
 
 int main();
-double sigmoid(double d);
-double sigmoid_prime(double d);
+float sigmoid(float d);
+float sigmoid_prime(float d);
 void forward(int nb_letter,
-		double **input,
-		double **hidden,
-		double **output,
-		double **w_input,
-		double **w_output,
-		double *b_input,
-		double *b_output);
+		float **input,
+		float **hidden,
+		float **output,
+		float **w_input,
+		float **w_output,
+		float *b_input,
+		float *b_output);
 void backward(int nb_letter,
-		double **w_output,
-		double **input,
-		double **hidden,
-		double **output,
-		double **expected_output,
-		double **output_error,
-		double **dw_output,
-		double *db_output,
-		double **hidden_error,
-		double **dw_input,
-		double *db_input);
-void update(double **w_input,
-		double **w_output,
-		double *b_input,
-		double *b_output,
-		double **dw_output,
-		double *db_output,
-		double **dw_input,
-		double *db_input,
-        double learning_rate);
+		float **w_output,
+		float **input,
+		float **hidden,
+		float **output,
+		float **expected_output,
+		float **output_error,
+		float **dw_output,
+		float *db_output,
+		float **hidden_error,
+		float **dw_input,
+		float *db_input);
+void update(float **w_input,
+		float **w_output,
+		float *b_input,
+		float *b_output,
+		float **dw_output,
+		float *db_output,
+		float **dw_input,
+		float *db_input,
+        float learning_rate);
 void train(int nb_letter,
 		char *anna_result,
-		double** w_input,
-		double** w_output,
-		double* b_input,
-		double* b_output);
-double log_loss(int nb_letter,
-        double **expected_output,
-        double **output);
-void poc_matrix_product_t(size_t row1, size_t col1, double **mat1,
-        size_t row2, size_t col2, double **mat2, double **res,
+		float** w_input,
+		float** w_output,
+		float* b_input,
+		float* b_output);
+float log_loss(int nb_letter,
+        float **expected_output,
+        float **output);
+void poc_matrix_product_t(size_t row1, size_t col1, float **mat1,
+        size_t row2, size_t col2, float **mat2, float **res,
 		char trans_nbr);
-void poc_matrix_product(size_t row1, size_t col1, double **mat1,
-		size_t row2, size_t col2, double **mat2, double **res);
+void poc_matrix_product(size_t row1, size_t col1, float **mat1,
+		size_t row2, size_t col2, float **mat2, float **res);
 
 #endif
