+++ HideWordSolver/ANNA/convolution.c
@@ -1,10 +1,10 @@
 #include "neural_network.h"
 
 void convolution(int nb_letter,
-                double **input,  // image_size^2 x nb_letter
-                double **output, // image_size^2 x nb_letter
-                double *filter1, // filter_size x filter_size
-                double *filter2) // filter_size x filter_size
+                float **input,  // image_size^2 x nb_letter
+                float **output, // image_size^2 x nb_letter
+                float *filter1, // filter_size x filter_size
+                float *filter2) // filter_size x filter_size
 {
     // Applying the first filter
     for (int i = 0; i < nb_letter; i++)
@@ -13,7 +13,7 @@ void convolution(int nb_letter,
         {
             for (size_t k = 0; k < image_size - filter_size + 1; k++)
             {
-                double sum = 0;
+                float sum = 0;
                 for (size_t l = 0; l < filter_size * filter_size; l++)
                 {
                     sum += input[(j * image_size + k) + l % filter_size +
@@ -32,7 +32,7 @@ void convolution(int nb_letter,
         {
             for (size_t k = 0; k < image_size - filter_size + 1; k++)
             {
-                double sum = 0;
+                float sum = 0;
                 for (size_t l = 0; l < filter_size * filter_size; l++)
                 {
                     sum += input[(j * image_size + k) + l % filter_size +
@@ -51,7 +51,7 @@ void convolution(int nb_letter,
         {
             for (size_t k = 0; k < image_size - filter_size + 1; k++)
             {
-                double sum = 0;
+                float sum = 0;
                 for (size_t l = 0; l < filter_size * filter_size; l++)
                 {
                     sum += input[(j * image_size + k) + l % filter_size +
