+++ HideWordSolver/ANNA/layers.c
@@ -1,21 +1,21 @@
 #include "neural_network.h"
 
 void layers(size_t nb_letter,
-		double **input,
-		double **output)
+		float **input,
+		float **output)
 {
-	double **temp_output = malloc(sizeof(double*) * image_size * image_size);
+	float **temp_output = malloc(sizeof(float*) * image_size * image_size);
 	for (size_t i = 0; i < image_size * image_size; i++)
 	{
-		temp_output[i] = malloc(sizeof(double) * nb_letter);
+		temp_output[i] = malloc(sizeof(float) * nb_letter);
 	}
 
-	double div = (double)1/(double)9;
+	float div = (float)1/(float)9;
 
-	double filter1[] = {div, div, div, 
+	float filter1[] = {div, div, div, 
 						div, div, div,
 						div, div, div};
-	double filter2[] = {0, -1, 0,
+	float filter2[] = {0, -1, 0,
 						-1, 5, -1,
 						0, -1, 0};
