+++ HideWordSolver/ANNA/predict.c
@@ -2,14 +2,14 @@
 
 void predict(int nb_letter,
 		char *anna_result,
-		double **input,
-        double **hidden,
-		double **output,
-		double **w_input,
-		double **w_output,
-		double *b_input,
-		double *b_output,
-		double threads)
+		float **input,
+        float **hidden,
+		float **output,
+		float **w_input,
+		float **w_output,
+		float *b_input,
+		float *b_output,
+		float threads)
 {
     forward(nb_letter, input, hidden, output, w_input, w_output,
 		b_input, b_output, -1, threads);
