+++ HideWordSolver/ANNA/relu_prime.c
@@ -1,4 +1,4 @@
-double relu_prime(double d)
+float relu_prime(float d)
 {
     return d > 0 ? 1.0 : 0.0;
 }
